+++ primedev/util/printmaps.cpp
@@ -136,9 +136,9 @@ int, __fastcall, (const char *const cmdname, const char *const partial, char com
 	RefreshMapList();
 
 	// use a custom autocomplete func for all map loading commands
-	const int cmdLength = strlen(cmdname);
+	const size_t cmdLength = strlen(cmdname);
 	const char* query = partial + cmdLength;
-	const int queryLength = strlen(query);
+	const size_t queryLength = strlen(query);
 
 	int numMaps = 0;
 	for (int i = 0; i < vMapList.size() && numMaps < COMMAND_COMPLETION_MAXITEMS; i++)
